refactor(vercelaisdk): remove unused history conversion
